worker-service/requirements.txt
qdrant-client==1.12.1
fastapi==0.115.12
uvicorn==0.33.0
python-multipart==0.0.20
insightface==0.7.3
opencv-python==4.11.0.86
numpy==1.24.4
onnxruntime-gpu==1.16.3
asyncpg==0.30.0
databases==0.9.0
boto3
python-dotenv
pytz
redis
Pillow
celery